requirements.txt
PyQt6==6.6.1
scapy==2.5.0
pyshark==0.6.0
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
pywin32==306; sys_platform == 'win32' 